requirements.txt
Django
fastapi
uvicorn
django-debug-toolbar
email-validator
flake8
pydantic
PyJWT
python-decouple
